Added Check for Successive Company IDs
We're now enforcing company_ids.json should have successive codes, to help prevent issues when changes are made down the line. This came after realising there is one Company ID that does not have an assignment in the official Bluetooth Numbers listing, but we added it to help the new test pass.